LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.clearRect.globalcomposite-expected.txt

2d.clearRect.globalcomposite
clearRect is not affected by globalCompositeOperation
Actual output:
Expected output:


PASS clearRect is not affected by globalCompositeOperation 
